fix: 학과 정보 타입 정의 위치 수정 (#56)
